+++ internal/component/faro/receiver/arguments.go
@@ -74,23 +74,35 @@ func (r *RateLimitingArguments) SetToDefault() {
 // SourceMapsArguments configures how app_agent_receiver will retrieve source
 // maps for transforming stack traces.
 type SourceMapsArguments struct {
-	Download                  bool                `alloy:"download,attr,optional"`
-	DownloadFromOrigins       []string            `alloy:"download_from_origins,attr,optional"`
-	DownloadTimeout           time.Duration       `alloy:"download_timeout,attr,optional"`
-	CacheMinimumTtl           time.Duration       `alloy:"cache_minimum_ttl,attr,optional"`
-	CacheErrorCleanupInterval time.Duration       `alloy:"cache_error_cleanup_interval,attr,optional"`
-	CacheCleanupCheckInterval time.Duration       `alloy:"cache_cleanup_check_interval,attr,optional"`
-	Locations                 []LocationArguments `alloy:"location,block,optional"`
+	Download            bool                `alloy:"download,attr,optional"`
+	DownloadFromOrigins []string            `alloy:"download_from_origins,attr,optional"`
+	DownloadTimeout     time.Duration       `alloy:"download_timeout,attr,optional"`
+	Cache               *CacheArguments     `alloy:"cache,block,optional"`
+	Locations           []LocationArguments `alloy:"location,block,optional"`
 }
 
 func (s *SourceMapsArguments) SetToDefault() {
 	*s = SourceMapsArguments{
-		Download:                  true,
-		DownloadFromOrigins:       []string{"*"},
-		DownloadTimeout:           time.Second,
-		CacheErrorCleanupInterval: time.Hour,
-		CacheMinimumTtl:           time.Duration(math.MaxInt64),
-		CacheCleanupCheckInterval: time.Second * 30,
+		Download:            true,
+		DownloadFromOrigins: []string{"*"},
+		DownloadTimeout:     time.Second,
+		Cache:               &CacheArguments{},
+	}
+	s.Cache.SetToDefault()
+}
+
+// CacheArguments configures sourcemap caching behavior.
+type CacheArguments struct {
+	Ttl                  time.Duration `alloy:"ttl,attr,optional"`
+	ErrorCleanupInterval time.Duration `alloy:"error_cleanup_interval,attr,optional"`
+	CleanupCheckInterval time.Duration `alloy:"cleanup_check_interval,attr,optional"`
+}
+
+func (c *CacheArguments) SetToDefault() {
+	*c = CacheArguments{
+		Ttl:                  time.Duration(math.MaxInt64),
+		ErrorCleanupInterval: time.Hour,
+		CleanupCheckInterval: time.Second * 30,
 	}
 }